+++ src/data/affiliation.ts
@@ -1,0 +1,4 @@
+export const affiliations = [
+    "british",
+    "french"
+] as const;