+++ src/pages/RulesPage/RulesPage.tsx
@@ -5,6 +5,29 @@ export function RulesPage() {
     return (
         <Page>
             <h1>Preparation</h1>
+            <p>
+                The game is played on a table. Preferably decorated with
+                scenery suitable for choosen scenario. Each player takes
+                control over an army (or at least a brigade) and commence
+                battle according to the game sequence.
+            </p>
+            <h1>Army preparation</h1>
+            <p>
+                Each army consists out of a general and one or more brigades.
+                Each brigade consists out of a commander and one or more units.
+                Each unit consists out of one or more bases/models.
+            </p>
+            <p>
+                In epic scale these are usually 60x20mm bases with multiple
+                models on them. A standard size unit is composed out of 4 of
+                such bases.
+            </p>
+            <p>
+                A commander or general is a single model on a base.
+            </p>
+            <p>
+                In epic scale a commander or general is placed on a 20x20mm base.
+            </p>
             <h1>Game sequence</h1>
             <p>
                 The game is played in turns. Within each turn all players